feat: add append_flags after redirection append every flag after a redirection to the previus command
