+++ include/append_flags.h
@@ -1,0 +1,20 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   append_flags.h                                     :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: rde-mour <rde-mour@student.42sp.org.br>    +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2024/03/21 10:59:22 by rde-mour          #+#    #+#             */
+/*   Updated: 2024/03/21 11:00:25 by rde-mour         ###   ########.org.br   */
+/*                                                                            */
+/* ************************************************************************** */
+
+#ifndef APPEND_FLAGS_H
+# define APPEND_FLAGS_H
+
+#include "ft_linkedlist.h"
+
+t_list	*append_flags(t_list *tokens);
+
+#endif
